Split renderer.py into focused core modules.
Extract settings, template loading, render context stamping, and component asset finalization so Renderer stays orchestration-only (~177 LOC).
