+++ pyjinhx/core/__init__.py
@@ -4,6 +4,9 @@ Core subsystem for pyjinhx.
 Modules:
 - ``base`` — ``BaseComponent`` and nesting wrappers
 - ``renderer`` — ``Renderer`` orchestration
+- ``renderer_settings`` — process-wide renderer defaults and factory cache
+- ``template_load`` — Jinja template discovery and loading
+- ``render_context`` — registry context injection and reactive stamping
 - ``render_assets`` — asset collection and injection helpers
 - ``tag_expand`` — PascalCase tag parsing and rendering
 - ``autodiscover`` — co-located component module import
